Created a new superclass of Line, handling child elements. This class is also intended as a superclass for Gedcom, to make one framework for iterating over child elements.
Some of the methods returning lists have been changed to
return iterators instead.  This is to prevent a caller
from modifying the contents of the list.

A new method has been introduced to return the first element with a
given tag, and this method is used instead of the iterators where
applicable.